+++ Lesson11/src/main/java/by/tms/clothes/model/Size.java
@@ -1,0 +1,24 @@
+package by.tms.clothes.model;
+
+public enum Size {
+    XXS(32) {
+        @Override
+        public String getDescription() {
+            return "Child size";
+        }
+    },
+    XS(34),
+    S(36),
+    M(38),
+    L(40);
+
+    private int euroSize;
+
+    Size(int euroSize) {
+        this.euroSize = euroSize;
+    }
+
+    public String getDescription() {
+        return "Adult clothes";
+    }
+}
