--- Lesson11/src/main/java/by/tms/homework2/model/Size.java
@@ -1,39 +1,0 @@
-package by.tms.homework2.model;
-
-// *
-////        * 1) Создать перечисление, содержащее размеры одежды (XXS, XS, S, M, L).
-////        * Перечисление содержит метод getDescription, возвращающий строку "Взрослый размер".
-////        * Переопределить метод getDescription - для константы XXS метод должен возвращать строку “Детский размер”.
-////        * Также перечисление должно содержать числовое значение euroSize(32, 34, 36, 38, 40), соответствующее каждому размеру.
-////        * Создать конструктор, принимающий на вход euroSize.
-////        * 2) Создать интерфейсы "Мужская Одежда" с методом "одетьМужчину" и "Женская Одежда" с методом "одетьЖенщину".
-////        * 3) Создать абстрактный класс Одежда, содержащий переменные - размер одежды, стоимость, цвет.
-////        * 4) Создать классы наследники Одежды - Футболка (реализует и интерфейсы "Мужская Одежда" и "Женская Одежда"),
-////        * Штаны (реализует интерфейсы "Мужская Одежда" и "Женская Одежда"), Юбка (реализует интерфейсы "Женская Одежда"),
-////        * Галстук (реализует интерфейсы "Мужская Одежда").
-////        * 5) Создать массив, содержащий все типы одежды. Создать класс Ателье, содержащий методы одетьЖенщину,
-////        * одетьМужчину, на вход которых будет поступать массив, содержащий все типы одежды.
-////        * Метод одетьЖенщину выводит на консоль всю информацию о женской одежде. То же самое для метода одетьМужчину.
-////        *
-public enum Size {
-    XXS(32) {
-        @Override
-        public String getDescription() {
-            return "Child size";
-        }
-    },
-    XS(34),
-    S(36),
-    M(38),
-    L(40);
-
-    private int euroSize;
-
-    Size(int euroSize) {
-        this.euroSize = euroSize;
-    }
-
-    public String getDescription() {
-        return "Adult clothes";
-    }
-}
