--- Lesson11/src/main/java/by/tms/homework2/model/Skirt.java
@@ -1,12 +1,0 @@
-package by.tms.homework2.model;
-
-import by.tms.homework2.service.WomanClothes;
-import lombok.experimental.SuperBuilder;
-
-@SuperBuilder
-public class Skirt extends Clothes implements WomanClothes {
-
-    @Override
-    public void dressWoman() {
-    }
-}
